app/requirements.txt
autopep8==1.5.7
pycodestyle==2.7.0
python-decouple==3.5
toml==0.10.2
